+++ app/src/main/AndroidManifest.xml
@@ -21,9 +21,9 @@
         </activity>
 
         <activity android:name=".landing.FtuActivity"/>
-        <activity android:name=".main.details.DetailsActivity"/>
+        <activity android:name=".tripdetail.DetailsActivity"/>
 
-        <activity android:name=".main.trips.CreateTripActivity"/>
+        <activity android:name=".newtrip.CreateTripActivity"/>
 
     </application>
