+++ app/src/main/java/ar/edu/itba/pam/travelapp/main/trips/TripsViewImpl.java
