+++ app/src/main/java/ar/edu/itba/pam/travelapp/tripdetail/DetailsActivity.java
@@ -1,4 +1,4 @@
-package ar.edu.itba.pam.travelapp.main.details;
+package ar.edu.itba.pam.travelapp.tripdetail;
 
 import android.os.AsyncTask;
 import android.os.Build;
