+++ app/src/main/java/ar/edu/itba/pam/travelapp/tripdetail/DetailsAdapter.java
@@ -1,4 +1,4 @@
-package ar.edu.itba.pam.travelapp.main.details;
+package ar.edu.itba.pam.travelapp.tripdetail;
 
 import android.os.Build;
 import android.view.LayoutInflater;
