+++ app/src/main/res/layout/activity_trip_details.xml
@@ -5,6 +5,6 @@
     android:id="@+id/trip_details"
     android:layout_height="match_parent"
     android:layout_width="match_parent"
-    tools:context=".main.details.DetailsActivity">
+    tools:context=".tripdetail.DetailsActivity">
 
 </androidx.recyclerview.widget.RecyclerView>
